pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "invest-klab-plugin"
# Allow setuptools_scm to derive the version dynamically from the git metadata
# Using setuptools_scm is not required, but we highly recommend it
dynamic = ["version"]  

# All direct python dependencies should be listed here. If a package is
# available from conda-forge, it can also be listed under the conda_dependencies.
dependencies = [
  "taskgraph",
  "pygeoprocessing",
  "natcap.invest>=3.17.0",
  "shapely>=2.0.7",
  "klab-client-py"
]

[tool.setuptools_scm]
# Use default configuration. A warning is issued if this section is absent.

[tool.natcap.invest]
# This is the plugin API version that this plugin uses.
# invest uses this to determine whether it can run this plugin.
# "v1" is the current and only version at this time
api_version = "v1"

# This is the importable package name of the plugin. It should match the name of the 
# package directory. It MUST begin with "invest". It is included here so that 
# natcap.invest knows how to import the plugin. 
# This is distinct from the project.name defined above (these names are often the same, 
# but don't have to be). Unlike the project.name, this name should not contain hyphens 
# or other special characters because it will be used in an import statement.
package_name = "invest_klab_plugin"

# List of conda dependencies to install into the plugin's conda environment. 
# Usually this includes python itself and any non-python dependencies
# like the GDAL binaries. Packages listed here will be installed into the
# plugin environment prior to the other dependencies, so you may wish to list
# packages here that are difficult to install with pip.
#
# Only the conda-forge channel is supported. These dependencies MUST be available on 
# conda-forge. We have disabled the default channel due to Anaconda's licensing.
#
# Python is installed into the plugin environment by default, but you may 
# include it here if you wish to specify a particular version
conda_dependencies = [
  "gdal<3.11",  # missing libgdal.so issue happened with gdal 3.11
  "geotiff>=1.7.3",  # earlier versions don't support space in CONDA_PREFIX
  "natcap.invest>=3.17.0",  # install from conda-forge rather than building wheel with pip
   "geopandas"
]
